Bad Apple (console with ASCII)/requirements.txt
opencv-python
Pillow
pygame
moviepy
fpstimer
inputvalidate